#13 Provide NotationModel and Notation Commands as separate plugins
