+++ plugins/org.eclipse.emfcloud.modelserver.glsp.integration/META-INF/MANIFEST.MF
@@ -17,8 +17,9 @@ Require-Bundle: org.eclipse.glsp.server,
  org.eclipse.elk.core;bundle-version="0.7.0",
  org.eclipse.elk.graph;bundle-version="0.7.0",
  org.eclipse.elk.alg.layered;bundle-version="0.7.0",
+ org.eclipse.emf.ecore,
  org.eclipse.emfcloud.modelserver.glsp.notation.model,
- org.eclipse.emf.ecore
+ org.eclipse.emfcloud.modelserver.glsp.notation.commands
 Bundle-ClassPath: .
 Export-Package: org.eclipse.emfcloud.modelserver.glsp,
  org.eclipse.emfcloud.modelserver.glsp.actions.handlers,
