+++ plugins/org.eclipse.emfcloud.modelserver.glsp.notation.commands/.checkstyle
@@ -1,0 +1,10 @@
+<fileset-config file-format-version="1.2.0" simple-config="false" sync-formatter="false">
+  <local-check-config name="EMF.cloud Checkstyle" location="https://raw.githubusercontent.com/eclipse-emfcloud/emfcloud/master/codestyle/org.eclipse.emfcloud.checkstyle/src/main/resources/emfcloud-checkstyle.xml" type="remote" description="The central checkstyle configuration for emfcloud">
+    <additional-data name="cache-props-file-location" value="EMFcloud Checkstyle_1618908540765_cache.properties"/>
+    <additional-data name="cache-file" value="true"/>
+    <additional-data name="cache-file-location" value="EMFcloud Checkstyle_1618908540765_cache.xml"/>
+  </local-check-config>
+  <fileset name="all" enabled="true" check-config-name="EMF.cloud Checkstyle" local="true">
+    <file-match-pattern match-pattern=".java$" include-pattern="true"/>
+  </fileset>
+</fileset-config>
