+++ plugins/org.eclipse.emfcloud.modelserver.glsp.notation.commands/META-INF/MANIFEST.MF
@@ -1,0 +1,24 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %pluginName
+Bundle-SymbolicName: org.eclipse.emfcloud.modelserver.glsp.notation.commands;singleton:=true
+Bundle-ActivationPolicy: lazy
+Bundle-Version: 0.7.0.qualifier
+Bundle-Vendor: %providerName
+Bundle-Localization: plugin
+Automatic-Module-Name: org.eclipse.emfcloud.modelserver.glsp.notation.commands
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Export-Package: org.eclipse.emfcloud.modelserver.glsp.notation.commands,
+ org.eclipse.emfcloud.modelserver.glsp.notation.commands.contribution,
+ org.eclipse.emfcloud.modelserver.glsp.notation.commands.util,
+ org.eclipse.emfcloud.modelserver.glsp.notation.epackage
+Require-Bundle: org.eclipse.emf.edit,
+ org.eclipse.emf.transaction,
+ org.eclipse.emfcloud.modelserver.common;bundle-version="0.7.0",
+ org.eclipse.emfcloud.modelserver.edit;bundle-version="0.7.0",
+ org.eclipse.emfcloud.modelserver.emf;bundle-version="0.7.0",
+ org.eclipse.emfcloud.modelserver.glsp.notation.model;bundle-version="0.7.0",
+ org.eclipse.emfcloud.modelserver.lib;bundle-version="0.7.0",
+ org.eclipse.glsp.graph;bundle-version="0.9.0",
+ org.eclipse.glsp.server;bundle-version="0.9.0"
+Bundle-ClassPath: .
