+++ plugins/org.eclipse.emfcloud.modelserver.glsp.notation.commands/build.properties
@@ -1,5 +1,4 @@
+source.. = src
 bin.includes = .,\
                META-INF/,\
-               plugin.xml,\
                plugin.properties
-source.. = src-gen
