+++ plugins/org.eclipse.emfcloud.modelserver.glsp.notation.commands/plugin.properties
@@ -1,0 +1,12 @@
+# Copyright (c) 2021 EclipseSource and others.
+#
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License v. 2.0 which is available at
+# https://www.eclipse.org/legal/epl-2.0, or the MIT License which is
+# available at https://opensource.org/licenses/MIT.
+#
+# SPDX-License-Identifier: EPL-2.0 OR MIT
+#
+
+pluginName = EMF.cloud Model Server Notation Commands Plugin
+providerName = EclipseSource
