+++ plugins/org.eclipse.emfcloud.modelserver.glsp.notation.commands/pom.xml
@@ -1,10 +1,8 @@
-
 <project xmlns="http://maven.apache.org/POM/4.0.0"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-
-	<artifactId>org.eclipse.emfcloud.modelserver.glsp.integration</artifactId>
+	<artifactId>org.eclipse.emfcloud.modelserver.glsp.notation.commands</artifactId>
 	<packaging>${package.type}</packaging>
 	<parent>
 		<groupId>org.eclipse.emfcloud.modelserver.glsp</groupId>
@@ -50,21 +48,27 @@
 			<activation>
 				<activeByDefault>false</activeByDefault>
 			</activation>
+
 			<dependencies>
-				<dependency>
-					<groupId>org.eclipse.emfcloud.modelserver</groupId>
-					<artifactId>org.eclipse.emfcloud.modelserver.emf</artifactId>
-					<version>${modelserver.version}</version>
-				</dependency>
 				<dependency>
 					<groupId>org.eclipse.emfcloud.modelserver.glsp</groupId>
 					<artifactId>org.eclipse.emfcloud.modelserver.glsp.notation.model</artifactId>
 					<version>${project.version}</version>
 				</dependency>
 				<dependency>
-					<groupId>org.eclipse.emfcloud.modelserver.glsp</groupId>
-					<artifactId>org.eclipse.emfcloud.modelserver.glsp.notation.commands</artifactId>
-					<version>${project.version}</version>
+					<groupId>org.eclipse.emfcloud.modelserver</groupId>
+					<artifactId>org.eclipse.emfcloud.modelserver.emf</artifactId>
+					<version>${modelserver.version}</version>
+				</dependency>
+				<dependency>
+					<groupId>org.eclipse.emfcloud.modelserver</groupId>
+					<artifactId>org.eclipse.emfcloud.modelserver.client</artifactId>
+					<version>${modelserver.version}</version>
+				</dependency>
+				<dependency>
+					<groupId>org.eclipse.glsp</groupId>
+					<artifactId>org.eclipse.glsp.graph</artifactId>
+					<version>${glsp.version}</version>
 				</dependency>
 				<dependency>
 					<groupId>org.eclipse.glsp</groupId>
@@ -72,9 +76,14 @@
 					<version>${glsp.version}</version>
 				</dependency>
 				<dependency>
-					<groupId>org.eclipse.glsp</groupId>
-					<artifactId>org.eclipse.glsp.layout</artifactId>
-					<version>${glsp.version}</version>
+					<groupId>org.eclipse.emf</groupId>
+					<artifactId>org.eclipse.emf.edit</artifactId>
+					<version>${emf.edit.version}</version>
+				</dependency>
+				<dependency>
+					<groupId>org.eclipse.emf</groupId>
+					<artifactId>transaction</artifactId>
+					<version>${emf.transaction.version}</version>
 				</dependency>
 			</dependencies>
