+++ plugins/org.eclipse.emfcloud.modelserver.glsp.notation.commands/src/org/eclipse/emfcloud/modelserver/glsp/notation/commands/contribution/ChangeBoundsCommandContribution.java
