+++ plugins/org.eclipse.emfcloud.modelserver.glsp.notation.model/META-INF/MANIFEST.MF
@@ -10,18 +10,7 @@ Automatic-Module-Name: org.eclipse.emfcloud.modelserver.glsp.notation.model
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Export-Package: org.eclipse.emfcloud.modelserver.glsp.notation,
  org.eclipse.emfcloud.modelserver.glsp.notation.impl,
- org.eclipse.emfcloud.modelserver.glsp.notation.util,
- org.eclipse.emfcloud.modelserver.glsp.notation.model,
- org.eclipse.emfcloud.modelserver.glsp.notation.commands,
- org.eclipse.emfcloud.modelserver.glsp.notation.commands.contribution,
- org.eclipse.emfcloud.modelserver.glsp.notation.commands.util
+ org.eclipse.emfcloud.modelserver.glsp.notation.util
 Require-Bundle: org.eclipse.emf.ecore,
- org.eclipse.glsp.graph;bundle-version="0.9.0",
- org.eclipse.glsp.server;bundle-version="0.9.0",
- org.eclipse.emfcloud.modelserver.common;bundle-version="0.7.0",
- org.eclipse.emfcloud.modelserver.edit;bundle-version="0.7.0",
- org.eclipse.emfcloud.modelserver.emf;bundle-version="0.7.0",
- org.eclipse.emfcloud.modelserver.lib;bundle-version="0.7.0",
- org.eclipse.emf.edit,
- org.eclipse.emf.transaction
+ org.eclipse.glsp.graph;bundle-version="0.9.0"
 Bundle-ClassPath: .
