+++ releng/org.eclipse.emfcloud.modelserver.glsp.feature/feature.xml
@@ -48,4 +48,11 @@
          version="0.0.0"
          unpack="false"/>
 
+   <plugin
+         id="org.eclipse.emfcloud.modelserver.glsp.notation.commands"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
 </feature>
